avancement mode pas a pas

+++ src/core/Case.cpp
@@ -20,7 +20,7 @@ unsigned char Case::getVal () const {
 
 void Case::setVal (unsigned char nb) {
     val=nb;
-    if (etat == 3) {
+    if ((int)etat == 3) {
         etat = 0;
     }
 }
